+++ src/main/java/com/supermartijn642/wormhole/portal/screen/PortalTargetScreen.java
@@ -52,7 +52,7 @@ public class PortalTargetScreen extends PortalGroupScreen {
     private static final ResourceLocation SEPARATOR = getTexture("select_target_screen/separator");
 
     private static ResourceLocation getTexture(String name){
-        return new ResourceLocation("wormhole", "textures/gui/" + name + ".png");
+        return ResourceLocation.fromNamespaceAndPath("wormhole", "textures/gui/" + name + ".png");
     }
 
     private static final int WIDTH = 240, HEIGHT = 185;