+++ changelog.md
@@ -1,2 +1,2 @@
 ### Wormhole 1.1.16
-- Initial release of Wormhole for Minecraft 1.20.5 & 1.20.6
+- Initial release of Wormhole for Minecraft 1.21 & 1.21.1